--- extensions/apps/extensions/src/components/pages/extension-submit-page/index.ts
@@ -1,4 +1,0 @@
-export * from './main-page';
-export * from './step1-page';
-export * from './step2-page';
-export * from './step3-page';
